[bcr-wallet-core] move pocket to a new module
